+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.6.0
+
+- Desktop notifications.
+- Ability to toggle notifications and sound.
+
 ## 0.5.0
 
 - Fix showing a tray icon.